flask_app/requirements.txt
llama-index==0.10.68
docx2txt==0.8
openpyxl==3.1.5
Flask==3.0.3
redisvl==0.1.3
llama-index-vector-stores-redis==0.2.1
python-dotenv==1.0.0